Constructure.java
The constructor is a special method that has the same name as its same class name. Default Constructor is a constructor that has no arguments or if it has parameters it will have default values. Create a class Main with the following member variables int length int breadth int height int radius In Main class, get the details of the values of all variables and pass it to the shape class Create a class Shape with the following member variables int length int breadth int height int radius The values are passed from the Main class as parameters to the constructors in Shape class. In Shape class, create four constructors. The first constructor is the default constructor.Shape(){} The second constructor is for the cube, which receives length, breadth, and height as arguments. Shape(int lenght,int breadth,int height){constructor body} The third constructor is for the sphere, which receives radius as arguments. Shape(int radius){constructor body} The fourth constructor is for cylinder, which receives radius and height as arguments. Shape(int radius, int height){Constructor body} Include three methods in Shape class namely public float calculateAreaOfCylinder() public float calculateAreaOfSphere() public Integer calculateAreaOfCube() to calculate the area of each shapes.

Sample Input and Output 1 :

1.Cube

2.Sphere

3.Cylinder

Enter the choice

1

length

10

breadth

20

height

30

Area of Cube is 6000



Sample Input and Output 2:

1.Cube

2.Sphere

3.Cylinder

Enter the choice

2

radius

10

Area of Sphere is 4186.67



Sample Input and Output 3:

1.Cube

2.Sphere

3.Cylinder

Enter the choice

3

radius

10

height

5

Area of Cylinders is 1570.0